+++ src/store/projectStore.ts
@@ -1,6 +1,6 @@
 import { create } from "zustand";
 
-import type { Project } from "@/types/project";
+import type { Project } from "@/types/Projects";
 
 type Store = {
     project: Project;
